pyzkds/requirements.txt
aiohttp==3.9.5
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.12.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
azure-cognitiveservices-speech==1.37.0
billiard==4.2.0
blinker==1.6.3
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
daphne==3.0.2
decorator==4.4.2
discord.py==2.3.2
distro==1.9.0
Django==4.0.5
django-cors-headers==4.3.0
django-cron==0.6.0
django-crontab==0.7.1
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.4.2
elasticsearch==7.7.0
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.14.0
Flask==2.2.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2024.3.1
googletrans==4.0.0rc1
greenlet==3.0.1
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hiredis==2.3.2
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==22.10.0
intel-openmp==2021.4.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.5
kombu==5.3.4
llvmlite==0.42.0
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.2
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==0.6.2
mtranslate==1.8
multidict==6.0.5
mysql-connector-python==8.2.0
mysqlclient==2.2.0
Naked==0.1.32
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy>=1.22.4,<1.27
pandas-stubs<2.2.0
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==9.5.0
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.20.0
pydantic==2.4.2
pydantic-core==2.10.1
pydub==0.25.1
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pysnowflake==0.1.3
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qiniu==7.12.0
redis==3.5.3
regex==2024.4.28
requests==2.31.0
rfc3986==1.5.0
ronglian-sms-sdk==1.0
rsa==4.9
service-identity==23.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
stopwords==1.0.1
sympy==1.12
tbb==2021.12.0
tencentcloud-sdk-python==3.0.1130
tiktoken==0.6.0
#torch==2.3.0
tornado==6.4
tqdm==4.66.4
twisted==23.10.0
#twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
wechatpayv3==1.2.42
werkzeug==3.0.0
whisper==1.1.10
yarl==1.9.4
zhconv==1.4.3
zipp==3.17.0
zope.interface==6.1
pyspark==3.5.3
thrift==0.21.0
thrift-sasl==0.4.3
PyHive==0.7.0
hdfs==2.7.3
scikit-learn==1.5.2
ultralytics==8.0.196

